Connected the app to firestore
