+++ src/components/ItemsList.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import firebase from './firebase';
+import firebase from '../lib/firebase';
 
 function useItems() {
   const [items, setItems] = useState([]);
